Moves the debug support into RMCdefines.h (to use it from rmc thread too). Removes kernel offset definitions in rmc.mod.c since it is generated
