+++ libsonuma/RMCdefines.h
@@ -85,5 +85,23 @@ typedef struct qp_info {
   int this_nid;
 } qp_info_t;
 
+// debug entry for printing a wq entry
+
+// Pure C implementation to return str rep. of WQ entry
+// FIXME: assumes buffer has enough space (please buffer-overflow attack this!)
+int stringify_wq_entry(wq_entry_t* entry,char* buf)
+{
+    return sprintf(buf,
+            "{ Operation = %c,"
+            " SR = %u,"
+            " Valid = %u,"
+            " LBuf_Addr = %lx,"
+            " LBuf_Offset = %lx,"
+            " Node ID = %d,"
+            " CTlx Offset = %lx,"
+            " Read Length = %lx }\n"
+            , entry->op, entry->SR, entry->valid, entry->buf_addr, entry->buf_offset,
+            entry->nid, entry->offset, entry->length);
+}
 
 #endif /* H_RMC_DEFINES */
